DF-7374 start improve old ResourceCompressor into new DavaPak
Former-commit-id: 7b8ce520d59c19f616eef168ce490a74c64b532c

+++ Tools/DavaPak/Classes/MemoryFile.cpp
@@ -148,7 +148,7 @@ bool MemoryFile::Seek(uint32 position, uint32 seekType)
 
 	
 	if (pos < 0)return false;
-	if (pos > fileSize)return false;
+	if (pos > (int64)fileSize)return false;
 	
 	dataCurrent = dataStart + (uint32)pos;
 	return true;
